Image build system
Experimented with docker bake, but then realized that I don't need
docker to manage my build for me (team of 1!) so I pivoted to a basic
Makefile that builds the programs and copies them into a docker image.

This commit has a bunch of other stuff as well and I frankly don't
remember what it was all for.