Fix extra optional letters bug
Fixes a bug would cause extra optional letters to
be set in some puzzle configs.

Puzzle configuration is built by querying the
words database in a loop with letter
requirements. This loop makes use of a variable
that holds a bitmask representing the optional
letters of the puzzle. That variable is set
outside of the loop. The loop failed to reset
this variable on every iteration and so optional
letters would bleed into subsequent iterations.

This fix adds the needed code to reset this
variable.