Fully safe letter mask construction
Makes letter mask construction fully safe from panics
(hopefully).

Previous attempts at this attempted to use ranges and
`Rng::rand_range` to build up letter masks. The idea was
that these ranges would be reduced as letters were picked
(to prevent duplicates). Ranges were used to avoid heap
allocations from Vecs or HashSets. In the end this
approach was complex for two primary reasons:
- ranges don't have built-in support for set operations like
  union, substraction, etc.
- `Rng::rand_range` panics on empty ranges! (sensible;
  duh, but there's no `try_rand_range`!)

This revision changes the logic to use the bitmasks
themselves to confirm whether a new letter has already
been selected and does this in a loop.