Avoid panic in letter selection
Prevents panics when randomly generating optional
letters in puzzle config.

`Rng::random_range` will panic if the provided
range is empty. The ranges used in this function
may be empty if the required letter is `'a'` --
this would cause the "front" range to be `'a'..'a'`.

This change adds more conditionals to handle either
front or back ranges being empty.