Namespace API endpoints
Put all API endpoints under the `/api/` namespace.

This is primarily being done to accommodate trunk
settings. Trunk allows you to specify a public url
for your compiled assets, but then it roots the dev
server to that public url. It also complains if you
try to proxy to `/`. So, now we manually set the
`--public-url` flag in the Makefile, don't set it
in our trunk config and namespace all of our APIs
to `/api/`